requirements.txt
transformers
accelerate
python-dotenv
pandas
torch  # Optional if already installed
